+++ front-end/src/app/service/comment.service.ts
@@ -21,68 +21,131 @@ export class CommentService {
     this.token = this.authService.getAuthToken();
   }
 
-  // Get all the comments
+
+  /**
+   * Get all comments.
+   * @returns Observable<Comment[]> - The array of comments.
+   */
   getComments(): Observable<Comment[]> {
+    // Check if the user is not authenticated
+    if (!this.authService.isLoggedIn()) {
+      return throwError(() => new Error('User not authenticated'));
+    }
+
+    // Send a GET request to retrieve the comments
     return this.http.get<Comment[]>(this.apiUrl)
     .pipe(
       map((response: any) => {
         return response;
       }),
       catchError((error: HttpErrorResponse) => {
-        console.error(error);
-        return throwError(() => error);
+        console.log(error);
+        return throwError(() => new Error('An error occurred while getting the comments'));
       })
     )
   }
 
-  // Create Comment method - params comment
+
+  /**
+   * Create a comment.
+   * @param comment - The comment object to create.
+   * @returns Observable<Comment> - The created comment.
+   */
   addComment(comment: Object): Observable<Comment> {
+    // Convert the comment object to JSON string
     const commentToSend = JSON.stringify(comment)
     const httpOptions = this.getHttpOptions();
 
+    // Check if the user is not authenticated
+    if (!this.authService.isLoggedIn()) {
+      return throwError(() => new Error('User not authenticated'));
+    }
+
+    // Send a POST request to create the comment
     return this.http.post<Comment>(`${this.apiUrl}/create-comment`, {comment: commentToSend}, httpOptions)
     .pipe(
       map((response: Comment) => {
         return response;
       }),
       catchError(error => {
-        return throwError(() => error);
+        console.log(error);
+        return throwError(() => new Error('An error occurred while creating the comment'));
       })
     );
   }
 
-  // Edit comment - params updatedComment - commentId
+
+  /**
+   * Edit a comment.
+   * @param updatedComment - The updated comment object.
+   * @param commentId - The ID of the comment to edit.
+   * @returns Observable<Comment> - The edited comment.
+   */
   editComment(updatedComment: Object, commentId: string): Observable<Comment> {
+    // Convert the updated comment object to JSON string
     const commentToSend = JSON.stringify(updatedComment);
     const httpOptions = this.getHttpOptions();
 
+    // Check if the user is not authenticated
+    if (!this.authService.isLoggedIn()) {
+      return throwError(() => new Error('User not authenticated'));
+    }
+
+    // Check if the commentId is invalid
+    if (!commentId) {
+      return throwError(() => new Error('Invalid commentId'));
+    }
+
+    // Send a PUT request to update the comment
     return this.http.put<Comment>(`${this.apiUrl}/${commentId}/update-comment`, {comment: commentToSend}, httpOptions)
     .pipe(
       map((response: any) => {
         return response;
       }),
       catchError(error => {
-        return throwError(() => error);
+        console.log(error);
+        return throwError(() => new Error('An error occurred while editing the comment'));
       })
     );
   }
 
-  // Delete Comment - params commentId
+
+  /**
+   * Delete a comment.
+   * @param commentId - The ID of the comment to delete.
+   * @returns Observable<Comment> - The deleted comment.
+   */
   deleteComment(commentId: string): Observable<Comment> {
     const httpOptions = this.getHttpOptions();
 
+    // Check if the user is not authenticated
+    if (!this.authService.isLoggedIn()) {
+      return throwError(() => new Error('User not authenticated'));
+    }
+
+    // Check if the commentId is invalid
+    if (!commentId) {
+      return throwError(() => new Error('Invalid commentId'));
+    }
+
+    // Send a DELETE request to delete the comment
     return this.http.delete<Comment>(`${this.apiUrl}/${commentId}/delete-comment`, httpOptions)
     .pipe(
       map((response: any) => {
         return response;
       }),
       catchError(error => {
-        return throwError(() => error);
+        console.log(error);
+        return throwError(() => new Error('An error occurred while deleting the comment'));
       })
     );
   }
 
-  // Declare the httpOptions
+
+  /**
+   * Get the HTTP options with authorization header.
+   * @returns any - The HTTP options object.
+   */
   getHttpOptions() {
     return {
       headers: new HttpHeaders({
